Add more complicated Enumerable w/ Enumerable tests
Add a test which has a List which contains elements which also
have a List of string.

From what I can tell, there's no way for a ValidationAttribute validator
to report back on all of the elements that failed validation within a
simple IEnumerable<string> list.  

At best, you might be able to write a ValidationAttribute validator which
reports back on the index of the first failure within the list.